feat(vue_oxlint_parser): extract lexer module and add script token collection
Address review feedback:

Lexer module: move all byte-level scanning helpers (current_byte, matches,
advance, skip_whitespace, peek_byte, etc.) from parser/mod.rs into a dedicated
parser/lexer/mod.rs as impl Parser<'a> blocks. Follows the same pattern as
vue_oxlint_jsx/parser/elements/. element.rs and attribute.rs stay focused on
V-node construction while lexer/ owns the low-level source navigation.

Token collection: vue-eslint-parser consumers require a Program.tokens array.
Switch oxc_parse_script to use TokensParserConfig so oxc_parser collects JS
tokens during script block parsing. Store them as script_tokens: Vec<Token> on
both Parser (internal) and VueSingleFileComponent (public). Token is oxc's
packed u128 (Copy) so the Vec<Token> allocation is cheap.

RFC update: add script_tokens to the VueSingleFileComponent struct definition,
document Token type and span semantics, and add a Module Layout section
describing the parser/lexer/ sub-module structure.
